requirements.txt
deepl==1.14.0
python-dotenv==0.21.1